print-engine/requirements.txt
fastapi
uvicorn
pydantic
Pillow
reportlab
PyMuPDF
python-multipart
requests
python-dotenv
websocket-client
numpy
anthropic
PyPDF2
